backend/requirements.txt
crewai[tools]>=0.28.8,<1.0
coqui-tts>=0.25.0
arxiv>=2.1.0
grobid-client-python>=0.0.5
fastapi>=0.95.2
uvicorn>=0.22.0
python-dotenv>=1.0.0
torch>=2.0.1 --extra-index-url https://download.pytorch.org/whl/cu118
pydantic>=2.0
boto3>=1.34.0
llama-cpp-python>=0.2.57
langchain-community>=0.0.28
